Rename tensoraddupdate.{h,cpp} -> tensor_add_update.{h,cpp}.
