+++ document/src/vespa/document/update/CMakeLists.txt
@@ -13,7 +13,7 @@ vespa_add_library(document_updates OBJECT
     mapvalueupdate.cpp
     removefieldpathupdate.cpp
     removevalueupdate.cpp
-    tensoraddupdate.cpp
+    tensor_add_update.cpp
     tensor_modify_update.cpp
     valueupdate.cpp
     DEPENDS
