+++ document/src/vespa/document/update/updates.h
@@ -10,6 +10,6 @@
 #include "clearvalueupdate.h"
 #include "mapvalueupdate.h"
 #include "removevalueupdate.h"
-#include "tensoraddupdate.h"
+#include "tensor_add_update.h"
 #include "tensor_modify_update.h"
