+++ searchcore/src/vespa/searchcore/proton/common/attribute_updater.cpp
@@ -14,7 +14,7 @@
 #include <vespa/document/update/clearvalueupdate.h>
 #include <vespa/document/update/mapvalueupdate.h>
 #include <vespa/document/update/removevalueupdate.h>
-#include <vespa/document/update/tensoraddupdate.h>
+#include <vespa/document/update/tensor_add_update.h>
 #include <vespa/document/update/tensor_modify_update.h>
 #include <vespa/eval/tensor/tensor.h>
 #include <vespa/searchlib/attribute/attributevector.hpp>
